# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Audit Wolf API
  description: |
    Comprehensive smart contract auditing platform API that combines AI-powered analysis 
    with traditional static analysis tools to provide detailed security assessments of Solidity contracts.
  version: 1.0.0
  contact:
    name: Audit Wolf Support
    email: support@auditwolf.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.auditwolf.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /reports/generate:
    post:
      tags:
        - Reports
      summary: Generate audit report
      description: Generate HTML and/or PDF report for completed audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReportRequest"
      responses:
        "200":
          description: Report generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportGenerationResponse"
        "400":
          description: Invalid request or audit not completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/{auditId}:
    get:
      tags:
        - Reports
      summary: Get report information
      description: Get information about existing reports for an audit
      parameters:
        - name: auditId
          in: path
          required: true
          description: Audit UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report information retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportInfoResponse"
        "404":
          description: Audit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/{auditId}/download/{format}:
    get:
      tags:
        - Reports
      summary: Download report file
      description: Download HTML or PDF report file
      parameters:
        - name: auditId
          in: path
          required: true
          description: Audit UUID
          schema:
            type: string
            format: uuid
        - name: format
          in: path
          required: true
          description: Report format
          schema:
            type: string
            enum: [html, pdf]
      responses:
        "200":
          description: File download
          content:
            text/html:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Report file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # Health Schemas
...
    GenerateReportRequest:
      type: object
      required:
        - auditId
        - format
      properties:
        auditId:
          type: string
          format: uuid
        format:
          type: string
          enum: [html, pdf, both]
        reportType:
          type: string
          enum: [standard, executive, detailed]
          default: standard
        includeSourceCode:
          type: boolean
          default: false
        customOptions:
          type: object
    ReportGenerationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            auditId:
              type: string
              format: uuid
            contractName:
              type: string
            generatedAt:
              type: string
              format: date-time
            html:
              type: object
              properties:
                available:
                  type: boolean
                downloadUrl:
                  type: string
            pdf:
              type: object
              properties:
                available:
                  type: boolean
                downloadUrl:
                  type: string
                size:
                  type: integer
                pages:
                  type: integer
    ReportInfoResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            auditId:
              type: string
              format: uuid
            hasReport:
              type: boolean
            reportGeneratedAt:
              type: string
              format: date-time
            files:
              type: object
              properties:
                html:
                  type: object
                  properties:
                    available:
                      type: boolean
                    size:
                      type: integer
                    downloadUrl:
                      type: string
                pdf:
                  type: object
                  properties:
                    available:
                      type: boolean
                    size:
                      type: integer
                    downloadUrl:
                      type: string

    # Multi-Blockchain Schemas
...